+++ async-http-client-backend/zio/src/main/scala/sttp/client/asynchttpclient/zio/ZioWebSocketHandler.scala
@@ -10,9 +10,6 @@ import zio.{DefaultRuntime, Queue, Runtime, Task, UIO}
 
 object ZioWebSocketHandler {
   private class ZioAsyncQueue[A](queue: Queue[A], runtime: Runtime[Any]) extends AsyncQueue[Task, A] {
-    override def clear(): Unit = {
-      val _ = runtime.unsafeRunToFuture(queue.takeAll)
-    }
     override def offer(t: A): Unit = {
       if (!runtime.unsafeRun(queue.offer(t))) {
         throw new WebSocketBufferFull()
